+++ datalad/support/annexrepo.py
@@ -15,6 +15,7 @@ For further information on git-annex see https://git-annex.branchable.com/.
 import json
 import logging
 import math
+import os
 import re
 import shlex
 
@@ -46,6 +47,7 @@ from datalad.support.external_versions import external_versions
 from datalad.support.external_versions import LooseVersion
 from datalad.support import ansi_colors
 from datalad.cmd import GitRunner
+from datalad.config import ConfigManager
 
 # imports from same module:
 from .gitrepo import GitRepo
@@ -79,7 +81,8 @@ class AnnexRepo(GitRepo):
     accepted either way.
     """
 
-    __slots__ = GitRepo.__slots__ + ['always_commit', '_batched', '_direct_mode', '_uuid']
+    __slots__ = GitRepo.__slots__ + ['always_commit', '_batched',
+                                     '_direct_mode', '_uuid']
 
     # Web remote has a hard-coded UUID we might (ab)use
     WEB_UUID = "00000000-0000-0000-0000-000000000001"
@@ -89,7 +92,7 @@ class AnnexRepo(GitRepo):
     git_annex_version = None
 
     def __init__(self, path, url=None, runner=None,
-                 direct=False, backend=None, always_commit=True, create=True,
+                 direct=None, backend=None, always_commit=True, create=True,
                  init=False, batch_size=None, version=None, description=None,
                  git_opts=None, annex_opts=None, annex_init_opts=None):
         """Creates representation of git-annex repository at `path`.
@@ -140,12 +143,12 @@ class AnnexRepo(GitRepo):
         # initialize
         self._uuid = None
 
-        if git_opts or annex_opts or annex_init_opts:
-            lgr.warning("TODO: options passed to git, git-annex and/or "
+        if annex_opts or annex_init_opts:
+            lgr.warning("TODO: options passed to git-annex and/or "
                         "git-annex-init are currently ignored.\n"
                         "options received:\n"
-                        "git: %s\ngit-annex: %s\ngit-annex-init: %s" %
-                        (git_opts, annex_opts, annex_init_opts))
+                        "git-annex: %s\ngit-annex-init: %s" %
+                        (annex_opts, annex_init_opts))
 
         fix_it = False
         try:
@@ -200,6 +203,18 @@ class AnnexRepo(GitRepo):
                             writer.release()
 
         self.always_commit = always_commit
+
+        # Note: Not sure yet, whether 'dataset=self' is appropriate, but we need
+        # to be able to read dataset's config in case we don't create a brand
+        # new one here.
+        cfg = ConfigManager(dataset=self)
+
+        if version is None:
+            try:
+                version = cfg["datalad.repo.version"]
+            except KeyError:
+                pass
+
         if fix_it:
             self._init(version=version, description=description)
             self.fsck()
@@ -220,10 +235,20 @@ class AnnexRepo(GitRepo):
             else:
                 raise RuntimeError("No annex found at %s." % self.path)
 
-        # only force direct mode; don't force indirect mode
+        # - only force direct mode; don't force indirect mode
+        # - parameter `direct` has priority over config
+        if direct is None:
+            direct = (create or init) and \
+                     cfg.getbool("datalad", "repo.direct", default=False)
         self._direct_mode = None  # we don't know yet
         if direct and not self.is_direct_mode():
-            self.set_direct_mode()
+            if self.repo.config_reader().get_value('annex', 'version') < 6:
+                lgr.debug("Switching to direct mode (%s)." % self)
+                self.set_direct_mode()
+            else:
+                # TODO: This may change to either not being a warning and/or
+                # to use 'git annex unlock' instead.
+                lgr.warning("direct mode not available for %s. Ignored." % self)
 
         # set default backend for future annex commands:
         # TODO: Should the backend option of __init__() also migrate
@@ -268,6 +293,11 @@ class AnnexRepo(GitRepo):
             # has no 2nd field in the key
             return None
         return int(size_str) if size_str.isdigit() else None
+
+    @normalize_path
+    def get_file_size(self, path):
+        fpath = opj(self.path, path)
+        return 0 if not exists(fpath) else os.stat(fpath).st_size
 
     @classmethod
     def is_valid_repo(cls, path, allow_noninitialized=False):
@@ -316,7 +346,7 @@ class AnnexRepo(GitRepo):
         return "<AnnexRepo path=%s (%s)>" % (self.path, type(self))
 
     def _run_annex_command(self, annex_cmd, git_options=None, annex_options=None,
-                           backend=None, **kwargs):
+                           backend=None, jobs=None, **kwargs):
         """Helper to run actual git-annex calls
 
         Unifies annex command calls.
@@ -333,6 +363,7 @@ class AnnexRepo(GitRepo):
             backend to be used by this command; Currently this can also be
             achieved by having an item '--backend=XXX' in annex_options.
             This may change.
+        jobs : int
         **kwargs
             these are passed as additional kwargs to datalad.cmd.Runner.run()
 
@@ -354,6 +385,9 @@ class AnnexRepo(GitRepo):
             cmd_list = ['git'] + git_options + ['annex']
         else:
             cmd_list = ['git-annex']
+        if jobs:
+            annex_options += ['-J%d' % jobs]
+
         cmd_list += [annex_cmd] + backend + debug + annex_options
 
         try:
@@ -437,6 +471,13 @@ class AnnexRepo(GitRepo):
         self._direct_mode = None
         assert(self.is_direct_mode() == enable_direct_mode)
 
+        if self.is_direct_mode():
+            # adjust git options for plain git calls on this repo:
+            # Note: Not sure yet, whether this solves the issue entirely or we
+            # still need 'annex proxy' in some cases ...
+            if 'core.bare=False' not in self._GIT_COMMON_OPTIONS:
+                self._GIT_COMMON_OPTIONS.extend(['-c', 'core.bare=False'])
+
     def _init(self, version=None, description=None):
         """Initializes an annex repository.
 
@@ -482,31 +523,8 @@ class AnnexRepo(GitRepo):
         # fetched
 
         if '--key' not in options:
-            lgr.debug("Determine what files need to be obtained")
-            # Let's figure out first which files/keys and of what size to download
-            expected_downloads = {}
-            fetch_files = []
-            keys_seen = set()
-            unknown_sizes = []  # unused atm
-
-            # for now just record total size, and
-            for j in self._run_annex_command_json(
-                'find', args=['--json', '--not', '--in', 'here'] + files
-                ):
-                key = j['key']
-                size = j.get('bytesize')
-                if key in keys_seen:
-                    # multiple files could point to the same key.  no need to
-                    # request multiple times
-                    continue
-                keys_seen.add(key)
-                assert j['file']
-                fetch_files.append(j['file'])
-                if size and size.isdigit():
-                    expected_downloads[key] = int(size)
-                else:
-                    expected_downloads[key] = None
-                    unknown_sizes.append(j['file'])
+            expected_downloads, fetch_files = self._get_expected_downloads(
+                files)
         else:
             fetch_files = files
             assert(len(files) == 1)
@@ -522,21 +540,10 @@ class AnnexRepo(GitRepo):
         # TODO:  check annex version and issue a one time warning if not
         # old enough for --json-progress
 
-        progress_indicators = ProcessAnnexProgressIndicators(
-            expected=expected_downloads,
-        )
-        run_kwargs = dict(
-            log_stdout=progress_indicators,
-            log_stderr='offline',  # False, # to avoid lock down
-            log_online=True
-        )
         # Without up to date annex, we would still report total! ;)
         if self.git_annex_version >= '6.20160923':
             # options  might be the '--key' which should go last
             options = ['--json-progress'] + options
-        if jobs:
-            options = ['-J%d' % jobs] + options
-
 
         # Note: Currently swallowing logs, due to the workaround to report files
         # not found, but don't fail and report about other files and use JSON,
@@ -556,17 +563,59 @@ class AnnexRepo(GitRepo):
         # failed to download
         with cm:
             results = self._run_annex_command_json(
-                'get', args=options + fetch_files, **run_kwargs)
+                'get', args=options + fetch_files,
+                jobs=jobs,
+                expected_entries=expected_downloads)
         results_list = list(results)
-        progress_indicators.finish()
         # TODO:  should we here compare fetch_files against result_list
         # and womit an exception of incomplete download????
         return results_list
 
+    def _get_expected_downloads(self, files):
+        """Given a list of files, figure out what to be downloaded
+
+        Parameters
+        ----------
+        files
+
+        Returns
+        -------
+        expected_downloads : dict
+          key -> size
+        fetch_files : list
+          files to be fetched
+        """
+        lgr.debug("Determine what files need to be obtained")
+        # Let's figure out first which files/keys and of what size to download
+        expected_downloads = {}
+        fetch_files = []
+        keys_seen = set()
+        unknown_sizes = []  # unused atm
+        # for now just record total size, and
+        for j in self._run_annex_command_json(
+                'find', args=['--json', '--not', '--in', 'here'] + files
+        ):
+            key = j['key']
+            size = j.get('bytesize')
+            if key in keys_seen:
+                # multiple files could point to the same key.  no need to
+                # request multiple times
+                continue
+            keys_seen.add(key)
+            assert j['file']
+            fetch_files.append(j['file'])
+            if size and size.isdigit():
+                expected_downloads[key] = int(size)
+            else:
+                expected_downloads[key] = None
+                unknown_sizes.append(j['file'])
+        return expected_downloads, fetch_files
 
     @normalize_paths
     def add(self, files, git=False, backend=None, options=None, commit=False,
-            msg=None, git_options=None, annex_options=None, _datalad_msg=False):
+            msg=None, dry_run=False,
+            jobs=None,
+            git_options=None, annex_options=None, _datalad_msg=False):
         """Add file(s) to the repository.
 
         Parameters
@@ -582,6 +631,9 @@ class AnnexRepo(GitRepo):
           the list of files that were added, is created by default.
         backend:
         options:
+        dry_run : bool, optional
+          Calls git add with --dry-run -N --ignore-missing, to just output list
+          of files to be added
 
         Returns
         -------
@@ -589,38 +641,56 @@ class AnnexRepo(GitRepo):
         """
 
         if git_options:
+            # TODO: note that below we would use 'add with --dry-run
+            # so passed here options might need to be passed into it??
             lgr.warning("git_options not yet implemented. Ignored.")
 
         if annex_options:
             lgr.warning("annex_options not yet implemented. Ignored.")
 
         options = options[:] if options else []
-
+        git_options = []
+        if dry_run:
+            git_options += ['--dry-run', '-N', '--ignore-missing']
         # Note: As long as we support direct mode, one should not call
         # super().add() directly. Once direct mode is gone, we might remove
         # `git` parameter and call GitRepo's add() instead.
         if git:
             # add to git instead of annex
             if self.is_direct_mode():
-                self.proxy(['git', 'add'] + options + files)
+                # TODO:  may be there should be a generic decorator to avoid
+                # duplication and just augment how git commands are called (i.e.
+                # with proxy
+                add_out = self.proxy(['git', 'add'] + options + git_options + files)
+                return_list = self._process_git_get_output(*add_out)
 
                 # Note/TODO:
                 # There was a reason to use the following instead of self.proxy:
                 #cmd_list = ['git', '-c', 'core.bare=false', 'add'] + options + \
                 #           files
                 #self.cmd_call_wrapper.run(cmd_list, expect_stderr=True)
-
-                # currently simulating return value, assuming success
-                # for all files:
-                return_list = [{u'file': f, u'success': True} for f in files]
-                # TODO: use options with git_add instead!
             else:
-                return_list = super(AnnexRepo, self).add(files)
+                return_list = super(AnnexRepo, self).add(files, git_options=git_options)
 
         else:
+            # Theoretically we could have done for git as well, if it could have
+            # been batched
+            # 1. Figure out what actually will be added
+            to_be_added_recs = self.add(files, git=True, dry_run=True)
+            # collect their sizes for the progressbar
+            expected_additions = {
+                rec['file']: self.get_file_size(rec['file'])
+                for rec in to_be_added_recs
+            }
             return_list = list(self._run_annex_command_json(
-                'add', args=options + files, backend=backend, expect_fail=True,
-                expect_stderr=True))
+                'add',
+                args=options + files,
+                backend=backend,
+                expect_fail=True,
+                jobs=jobs,
+                expected_entries=expected_additions,
+                expect_stderr=True
+            ))
 
         if commit:
             if msg is None:
@@ -655,68 +725,80 @@ class AnnexRepo(GitRepo):
         (stdout, stderr)
             output of the command call
         """
+        # TODO: We probably don't need it anymore
 
         if not self.is_direct_mode():
             lgr.warning("proxy() called in indirect mode: %s" % git_cmd)
             raise CommandNotAvailableError(cmd="git annex proxy",
                                            msg="Proxy doesn't make sense"
                                                " if not in direct mode.")
-        # Temporarily use shlex, until calls use lists for git_cmd
         return self._run_annex_command('proxy',
                                        annex_options=['--'] + git_cmd,
                                        **kwargs)
 
-    @normalize_path
-    def get_file_key(self, file_):
+    @normalize_paths
+    def get_file_key(self, files):
         """Get key of an annexed file.
 
         Parameters
         ----------
-        file_: str
-            file to look up
+        files: str or list
+            file(s) to look up
 
         Returns
         -------
-        str
-            keys used by git-annex for each of the files
+        str or list
+            keys used by git-annex for each of the files;
+            in case of a list an empty string is returned if there was no key
+            for that file
         """
 
-        cmd_str = 'git annex lookupkey %s' % file_  # have a string for messages
+        if len(files) > 1:
+            return self._batched.get('lookupkey',
+                                     git_options=self._GIT_COMMON_OPTIONS,
+                                     path=self.path)(files)
+        else:
+            files = files[0]
+            # single file
+            # keep current implementation
+            # TODO: This should change, but involves more RF'ing and an
+            # alternative regarding FileNotInAnnexError
+            cmd_str = 'git annex lookupkey %s' % files  # have a string for messages
 
-        try:
-            out, err = self._run_annex_command('lookupkey',
-                                               annex_options=[file_],
-                                               expect_fail=True)
-        except CommandError as e:
-            if e.code == 1:
-                if not exists(opj(self.path, file_)):
-                    raise IOError(e.code, "File not found.", file_)
-                # XXX you don't like me because I can be real slow!
-                elif file_ in self.get_indexed_files():
-                    # if we got here, the file is present and in git,
-                    # but not in the annex
-                    raise FileInGitError(cmd=cmd_str,
-                                         msg="File not in annex, but git: %s"
-                                             % file_,
-                                         filename=file_)
+            try:
+                out, err = self._run_annex_command('lookupkey',
+                                                   annex_options=[files],
+                                                   expect_fail=True)
+            except CommandError as e:
+                if e.code == 1:
+                    if not exists(opj(self.path, files)):
+                        raise IOError(e.code, "File not found.", files)
+                    # XXX you don't like me because I can be real slow!
+                    elif files in self.get_indexed_files():
+                        # if we got here, the file is present and in git,
+                        # but not in the annex
+                        raise FileInGitError(cmd=cmd_str,
+                                             msg="File not in annex, but git: %s"
+                                                 % files,
+                                             filename=files)
+                    else:
+                        raise FileNotInAnnexError(cmd=cmd_str,
+                                                  msg="File not in annex: %s"
+                                                      % files,
+                                                  filename=files)
                 else:
-                    raise FileNotInAnnexError(cmd=cmd_str,
-                                              msg="File not in annex: %s"
-                                                  % file_,
-                                              filename=file_)
-            else:
-                # Not sure, whether or not this can actually happen
-                raise e
+                    # Not sure, whether or not this can actually happen
+                    raise e
 
-        entries = out.rstrip(linesep).splitlines()
-        # filter out the ones which start with (: http://git-annex.branchable.com/bugs/lookupkey_started_to_spit_out___34__debug__34___messages_to_stdout/?updated
-        entries = list(filter(lambda x: not x.startswith('('), entries))
-        if len(entries) > 1:
-            lgr.warning("Got multiple entries in reply asking for a key of a file: %s"
-                        % (str(entries)))
-        elif not entries:
-            raise FileNotInAnnexError("Could not get a key for a file %s -- empty output" % file_)
-        return entries[0]
+            entries = out.rstrip(linesep).splitlines()
+            # filter out the ones which start with (: http://git-annex.branchable.com/bugs/lookupkey_started_to_spit_out___34__debug__34___messages_to_stdout/?updated
+            entries = list(filter(lambda x: not x.startswith('('), entries))
+            if len(entries) > 1:
+                lgr.warning("Got multiple entries in reply asking for a key of a file: %s"
+                            % (str(entries)))
+            elif not entries:
+                raise FileNotInAnnexError("Could not get a key for a file(s) %s -- empty output" % files)
+            return entries[0]
 
     @normalize_paths(map_filenames_back=True)
     def find(self, files, batch=False):
@@ -737,7 +819,7 @@ class AnnexRepo(GitRepo):
         """
         objects = []
         if batch:
-            objects = self._batched.get('find', path=self.path)(files)
+            objects = self._batched.get('find', git_options=self._GIT_COMMON_OPTIONS, path=self.path)(files)
         else:
             for f in files:
                 try:
@@ -920,7 +1002,7 @@ class AnnexRepo(GitRepo):
                 # Since backend will be critical for non-existing files
                 'addurl_to_file_backend:%s' % backend,
                 annex_cmd='addurl',
-                git_options=git_options,
+                git_options=self._GIT_COMMON_OPTIONS + git_options,
                 annex_options=options,  # --raw ?
                 path=self.path,
                 json=True
@@ -942,6 +1024,7 @@ class AnnexRepo(GitRepo):
             return out_json
 
     def add_urls(self, urls, options=None, backend=None, cwd=None,
+                 jobs=None,
                  git_options=None, annex_options=None):
         """Downloads each url to its own file, which is added to the annex.
 
@@ -1064,7 +1147,7 @@ class AnnexRepo(GitRepo):
         if not batch:
             json_objects = self._run_annex_command_json('dropkey', args=options + keys, expect_stderr=True)
         else:
-            json_objects = self._batched.get('dropkey', annex_options=options, json=True, path=self.path)(keys)
+            json_objects = self._batched.get('dropkey', git_options=self._GIT_COMMON_OPTIONS, annex_options=options, json=True, path=self.path)(keys)
         for j in json_objects:
             assert j.get('success', True)
 
@@ -1080,12 +1163,32 @@ class AnnexRepo(GitRepo):
             assert(remotes[self.WEB_UUID]['description'] == 'web')
         return remotes
 
-    def _run_annex_command_json(self, command, args=None, **kwargs):
+    def _run_annex_command_json(self, command, args=None, jobs=None,
+                                expected_entries=None, **kwargs):
         """Run an annex command with --json and load output results into a tuple of dicts
+
+        Parameters
+        ----------
+        expected_entries : dict, optional
+          If provided `filename/key: size` dictionary, will be used to create
+          ProcessAnnexProgressIndicators to display progress
         """
+        progress_indicators = None
         try:
+            if expected_entries:
+                progress_indicators = ProcessAnnexProgressIndicators(
+                    expected=expected_entries
+                )
+                kwargs = kwargs.copy()
+                kwargs.update(dict(
+                    log_stdout=progress_indicators,
+                    log_stderr='offline',  # False, # to avoid lock down
+                    log_online=True
+                ))
             # TODO: refactor to account for possible --batch ones
             annex_options = ['--json']
+            if jobs:
+                annex_options += ['-J%d' % jobs]
             if args:
                 annex_options += args
             out, err = self._run_annex_command(
@@ -1163,6 +1266,9 @@ class AnnexRepo(GitRepo):
             # failure in keys 'success' and 'note'
             if out is None:
                 raise e
+        finally:
+            if progress_indicators:
+                progress_indicators.finish()
 
         json_objects = (json.loads(line)
                         for line in out.splitlines() if line.startswith('{'))
@@ -1253,7 +1359,7 @@ class AnnexRepo(GitRepo):
         if not batch:
             json_objects = self._run_annex_command_json('info', args=options + files)
         else:
-            json_objects = self._batched.get('info', annex_options=options, json=True, path=self.path)(files)
+            json_objects = self._batched.get('info', git_options=self._GIT_COMMON_OPTIONS, annex_options=options, json=True, path=self.path)(files)
 
         # Some aggressive checks. ATM info can be requested only per file
         # json_objects is a generator, let's keep it that way
@@ -1406,7 +1512,7 @@ class AnnexRepo(GitRepo):
             except CommandError:
                 return ''
         else:
-            return self._batched.get('contentlocation', path=self.path)(key)
+            return self._batched.get('contentlocation', git_options=self._GIT_COMMON_OPTIONS, path=self.path)(key)
 
     @normalize_paths(serialize=True)
     def is_available(self, file_, remote=None, key=False, batch=False):
@@ -1451,7 +1557,7 @@ class AnnexRepo(GitRepo):
                 return False
         else:
             annex_cmd = ["checkpresentkey"] + ([remote] if remote else [])
-            out = self._batched.get(':'.join(annex_cmd), annex_cmd, path=self.path)(key_)
+            out = self._batched.get(':'.join(annex_cmd), annex_cmd, git_options=self._GIT_COMMON_OPTIONS, path=self.path)(key_)
             try:
                 return {
                     '': False,  # when remote is misspecified ... stderr carries the msg
@@ -1753,7 +1859,7 @@ class BatchedAnnex(object):
         # TODO -- should get all those options about --debug and --backend which are used/composed
         # in AnnexRepo class
         lgr.debug("Initiating a new process for %s" % repr(self))
-        cmd = ['git'] + AnnexRepo._GIT_COMMON_OPTIONS + self.git_options + \
+        cmd = ['git'] + self.git_options + \
               ['annex'] + self.annex_cmd + self.annex_options + ['--batch']  # , '--debug']
         lgr.log(5, "Command: %s" % cmd)
         # TODO: look into _run_annex_command  to support default options such as --debug