+++ datalad/support/gitrepo.py
@@ -41,6 +41,7 @@ from git.objects.blob import Blob
 from datalad import ssh_manager
 from datalad.cmd import Runner, GitRunner
 from datalad.dochelpers import exc_str
+from datalad.utils import assure_list
 from datalad.utils import optional_args
 from datalad.utils import on_windows
 from datalad.utils import getpwd
@@ -70,7 +71,8 @@ default_git_odbt = gitpy.GitCmdObjectDB
 # log Exceptions from git commands.
 
 
-# TODO: ignore leading and/or trailing underscore to allow for python-reserved words
+# TODO: ignore leading and/or trailing underscore to allow for
+# python-reserved words
 @optional_args
 def kwargs_to_options(func, split_single_char_options=True,
                       target_kw='options'):
@@ -374,13 +376,8 @@ class GitRepo(object):
     overridden accidentally by AnnexRepo.
 
     """
-    __slots__ = ['path', 'repo', 'cmd_call_wrapper']
 
-    # Disable automatic garbage and autopacking
-    _GIT_COMMON_OPTIONS = ['-c', 'receive.autogc=0', '-c', 'gc.auto=0']
-    # actually no need with default GitPython db backend not in memory
-    # default_git_odbt but still allows for faster testing etc.
-    # May be eventually we would make it switchable _GIT_COMMON_OPTIONS = []
+    __slots__ = ['path', 'repo', 'cmd_call_wrapper', '_GIT_COMMON_OPTIONS']
 
     def __init__(self, path, url=None, runner=None, create=True,
                  git_opts=None, **kwargs):
@@ -423,9 +420,16 @@ class GitRepo(object):
 
         """
 
-        if git_opts:
-            lgr.warning("TODO: options passed to git are currently ignored.\n"
-                        "options received: %s" % git_opts)
+        # Disable automatic garbage and autopacking
+        self._GIT_COMMON_OPTIONS = ['-c', 'receive.autogc=0', '-c', 'gc.auto=0']
+        # actually no need with default GitPython db backend not in memory
+        # default_git_odbt but still allows for faster testing etc.
+        # May be eventually we would make it switchable _GIT_COMMON_OPTIONS = []
+
+        if git_opts is None:
+            git_opts = {}
+        if kwargs:
+            git_opts.update(kwargs)
 
         # Sanity check for argument `path`:
         # raise if we cannot deal with `path` at all or
@@ -464,11 +468,14 @@ class GitRepo(object):
 
         if create and not GitRepo.is_valid_repo(path):
             try:
-                lgr.debug("Initialize empty Git repository at {0}".format(path))
+                lgr.debug(
+                    "Initialize empty Git repository at '%s'%s",
+                    path,
+                    ' %s' % git_opts if git_opts else '')
                 self.repo = self.cmd_call_wrapper(gitpy.Repo.init, path,
                                                   mkdir=True,
                                                   odbt=default_git_odbt,
-                                                  **kwargs)
+                                                  **git_opts)
             except GitCommandError as e:
                 lgr.error(exc_str(e))
                 raise
@@ -482,6 +489,12 @@ class GitRepo(object):
                         InvalidGitRepositoryError) as e:
                     lgr.error("%s: %s" % (type(e), str(e)))
                     raise
+
+        # inject git options into GitPython's git call wrapper:
+        # Note: `None` currently can happen, when Runner's protocol prevents
+        # calls above from being actually executed (DryRunProtocol)
+        if self.repo is not None:
+            self.repo.git._persistent_git_options = self._GIT_COMMON_OPTIONS
 
     def clone(self, url, path):
         """Clone url into path
@@ -621,7 +634,7 @@ class GitRepo(object):
         Parameters
         ----------
         files: list
-            list of paths to add
+          list of paths to add
         commit: bool
           whether or not to directly commit
         msg: str
@@ -632,18 +645,21 @@ class GitRepo(object):
           has to be always true.
         """
 
-        if git_options:
-            lgr.warning("git_options not yet implemented. Ignored.")
-
         # needs to be True - see docstring:
         assert(git)
 
         files = _remove_empty_items(files)
+        out = []
+
         if files:
             try:
-
-                self._git_custom_command(files, ['git', 'add'])
-
+                # without --verbose git 2.9.3  add does not return anything
+                add_out = self._git_custom_command(
+                    files,
+                    ['git', 'add'] + assure_list(git_options) + ['--verbose']
+                )
+                # get all the entries
+                out = self._process_git_get_output(*add_out)
                 # Note: as opposed to git cmdline, force is True by default in
                 #       gitpython, which would lead to add things, that are
                 #       ignored or excluded otherwise
@@ -678,7 +694,17 @@ class GitRepo(object):
         # currently simulating similar return value, assuming success
         # for all files:
         # TODO: Make return values consistent across both *Repo classes!
-        return [{u'file': f, u'success': True} for f in files]
+        return out
+
+    @staticmethod
+    def _process_git_get_output(stdout, stderr=None):
+        """Given both outputs (stderr is ignored atm) of git add - process it
+
+        Primarily to centralize handling in both indirect annex and direct
+        modes when ran through proxy
+        """
+        return [{u'file': f, u'success': True}
+                for f in re.findall("'(.*)'[\n$]", stdout)]
 
     @normalize_paths(match_return_type=False)
     def remove(self, files, **kwargs):
@@ -789,7 +815,13 @@ class GitRepo(object):
         # TODO: support not only a branch but any treeish
         #       Note: repo.tree(treeish).hexsha
         if branch is None:
-            return self.repo.active_branch.object.hexsha
+            try:
+                return self.repo.active_branch.object.hexsha
+            except ValueError as exc:
+                if 'does not exist' in str(exc):
+                    return None
+                raise
+
         for b in self.repo.branches:
             if b.name == branch:
                 return b.object.hexsha